[imp] - styles moved to catalog and typography to mixin
