+++ src/mainLib.js
@@ -7,7 +7,7 @@ import BBadge from "./components/b-badge";
 import BInput from "./components/b-input";
 import BSelect from "./components/b-select";
 import BDatepicker from "./components/b-datepicker";
-import "./base.scss";
+import "./assets/styles/base.scss";
 
 export default {
   install(Vue) {
